+++ extensions/Splitwise/src/hooks/useCurrentUser.tsx
@@ -1,13 +1,18 @@
 import { useFetch } from "@raycast/utils";
 
 import { GetUser } from "../types/get_user.types"; // Types
-import { HEADER } from "./userPreferences";
+import { useOAuth } from "./useOAuth";
 
 export function GetCurrentUser() {
+  const tokenSet = useOAuth();
+
   const { data, error } = useFetch<GetUser>(`https://secure.splitwise.com/api/v3.0/get_current_user`, {
     method: "GET",
-    ...HEADER,
+    headers: {
+      Authorization: `Bearer ${tokenSet?.accessToken}`,
+    },
     keepPreviousData: true,
+    execute: !!tokenSet,
   });
 
   const currentUser = data?.user;